+++ src/para.cpp
@@ -157,7 +157,7 @@ void free_state(State st) {
     return;
 }
 
-int local_set(int data[], int row, int color) {
+int local_set(int8_t data[], int row, int color) {
         if (row < 0 || row >= global_height) 
             return OUTOFBOUNDS;
 
@@ -298,7 +298,7 @@ bool solve(Puzzle p, Solution sol) {
                 /* #pragma omp for */  
                 /* for (int i = 0; i < height; i++){ */
                     int size = solv->row_sizes[i];
-                    int local[width]; //get local copy
+                    int8_t local[width]; //get local copy
                     bool lconflict = false;
                     bool lprogress = false; 
                     for(int j = 0; j < width; j++){
@@ -688,7 +688,7 @@ bool solve(Puzzle p, Solution sol) {
                     if (!col_solved(i, solu, p, solv)) {
                 //#pragma omp for  
                 /* for (int i = 0; i < width; i++) { */
-                    int local_col[height]; //get local copy
+                    int8_t local_col[height]; //get local copy
                     bool lconflict = false;
                     bool lprogress = false; 
                     for(int j = 0; j < height; j++){
